extracted common.py
