+++ depth_and_gsam.py
@@ -12,8 +12,10 @@ import numpy as np
 import skimage
 import matplotlib
 
+import common
 import zedhelper.handmark
 import zedhelper.util
+from common import parse_args, resize_image
 from zedhelper import predefined
 
 import gsam_module
@@ -22,51 +24,6 @@ import inspect
 
 MAX_ABS_DEPTH, MIN_ABS_DEPTH = 0.0, 2.0  # [m]
 
-def parse_args(init):
-    if len(opt.input_svo_file) > 0 and opt.input_svo_file.endswith(".svo"):
-        init.set_from_svo_file(opt.input_svo_file)
-        print("[Sample] Using SVO File input: {0}".format(opt.input_svo_file))
-    elif len(opt.ip_address) > 0:
-        ip_str = opt.ip_address
-        if (
-            ip_str.replace(":", "").replace(".", "").isdigit()
-            and len(ip_str.split(".")) == 4
-            and len(ip_str.split(":")) == 2
-        ):
-            init.set_from_stream(ip_str.split(":")[0], int(ip_str.split(":")[1]))
-            print("[Sample] Using Stream input, IP : ", ip_str)
-        elif ip_str.replace(":", "").replace(".", "").isdigit() and len(ip_str.split(".")) == 4:
-            init.set_from_stream(ip_str)
-            print("[Sample] Using Stream input, IP : ", ip_str)
-        else:
-            print("Unvalid IP format. Using live stream")
-    if "HD2K" in opt.resolution:
-        init.camera_resolution = sl.RESOLUTION.HD2K
-        print("[Sample] Using Camera in resolution HD2K")
-    elif "HD1200" in opt.resolution:
-        init.camera_resolution = sl.RESOLUTION.HD1200
-        print("[Sample] Using Camera in resolution HD1200")
-    elif "HD1080" in opt.resolution:
-        init.camera_resolution = sl.RESOLUTION.HD1080
-        print("[Sample] Using Camera in resolution HD1080")
-    elif "HD720" in opt.resolution:
-        init.camera_resolution = sl.RESOLUTION.HD720
-        print("[Sample] Using Camera in resolution HD720")
-    elif "SVGA" in opt.resolution:
-        init.camera_resolution = sl.RESOLUTION.SVGA
-        print("[Sample] Using Camera in resolution SVGA")
-    elif "VGA" in opt.resolution:
-        init.camera_resolution = sl.RESOLUTION.VGA
-        print("[Sample] Using Camera in resolution VGA")
-    elif len(opt.resolution) > 0:
-        print("[Sample] No valid resolution entered. Using default")
-    else:
-        print("[Sample] Using default resolution")
-
-
-def resize_image(image: np.ndarray, rate: float) -> np.ndarray:
-    H, W = image.shape[:2]
-    return cv2.resize(image, (int(W * rate), int(H * rate)))
 
 def points_by_segmentation(points: np.ndarray, segmentation_image: np.ndarray):
     """
@@ -152,7 +109,7 @@ def main(opt):
 
     init_params = predefined.InitParameters()
 
-    parse_args(init_params)
+    parse_args(opt, init_params)
 
     init_params.depth_mode = sl.DEPTH_MODE.ULTRA
     # init_params.depth_mode = sl.DEPTH_MODE.NEURAL2
